COZMO-11128 Some animations play differently on android/kindle vs iOS (#4073)
- When loaded in from flat buffers or json, the radius value for TURN_IN_PLACE is converted from string -> float -> uint16_t -> int16_t. For negative radius, this ideally looks like: "-12.0" -> -12.0 -> 65524 -> -12. But the last conversion (from u16 to s16, which happens implicitly) is actually undefined behavior, and was behaving different on macOS/iOS and Android.
- Removing the unsigned cast

+++ basestation/src/anki/cozmo/basestation/keyframe.cpp
@@ -778,7 +778,7 @@ _streamMsg.colors[__LED_NAME__] = ENCODED_COLOR(color); } while(0)
 
       const std::string& radiusStr = bodyKeyframe->radius_mm()->str();
       if (has_any_digits(radiusStr)) {
-        _streamMsg.curvatureRadius_mm = (uint16_t) std::atof(radiusStr.c_str());
+        _streamMsg.curvatureRadius_mm = std::atoi(radiusStr.c_str());
         CheckTurnSpeed(animNameDebug);
       } else {
         Result lastResult = ProcessRadiusString(radiusStr, animNameDebug);
@@ -807,7 +807,7 @@ _streamMsg.colors[__LED_NAME__] = ENCODED_COLOR(color); } while(0)
           return lastResult;
         }
       } else {
-        _streamMsg.curvatureRadius_mm = (uint16_t)jsonRoot["radius_mm"].asInt();
+        _streamMsg.curvatureRadius_mm = jsonRoot["radius_mm"].asInt();
         CheckTurnSpeed(animNameDebug);
       }
